refactor: using-Direktiven und Namensräume bereinigt
Unnötige und redundante using-Direktiven in zahlreichen Dateien entfernt, Namensraum-Präfixe vereinheitlicht und Methodenaufrufe vereinfacht. PropertyChangedEventArgs und OnPropertyChanged-Aufrufe modernisiert. Typen wie Color und Point konsistent verwendet. Debug-Ausgaben und Eventhandler-Parameter vereinheitlicht. Verbesserte Lesbarkeit, Wartbarkeit und Konsistenz im gesamten Code.

+++ Backend/Protocol/Z21MessageParser.cs
@@ -171,7 +171,7 @@ public static class Z21MessageParser
 
             locoInfo = new LocoInfo
             {
-                Address = (int)address,
+                Address = address,
                 Speed = speed,
                 IsForward = forward,
                 SpeedSteps = speedSteps switch { 0 => 14, 2 => 28, 3 => 128, _ => 128 },
